cast to unsigned should work everywhere
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3393 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/lua/loslibext.c
@@ -639,7 +639,7 @@ static int ex_sleep(lua_State * L)
 #ifdef WIN32
     Sleep(1e3 * interval / units);
 #else                           /* assumes posix or bsd */
-    usleep((1e6 * interval / units));
+    usleep((unsigned)(1e6 * interval / units));
 #endif
     return 0;
 }
